A minor hack to rewrite the output of git log
While git log --format handles all formatting concerns up until a diff
and/or a diffstat, there is no way to add custom labels to those two
parts of the output.

This hack amends the output with a custom label for the diff, while
making sure that the `---` separator is not added after the message body,
allowing the last line of the --format to be the diffstat label.